+++ job_source_agent/providers/__init__.py
@@ -1,4 +1,6 @@
 from .base import AdapterResult, JobBoard, JobCandidate, JobQuery, ProviderAdapter
+from .greenhouse import GreenhouseAdapter
+from .registry import DEFAULT_PROVIDER_REGISTRY, ProviderRegistry, build_default_provider_registry
 
 __all__ = [
     "AdapterResult",
@@ -6,5 +8,8 @@ __all__ = [
     "JobCandidate",
     "JobQuery",
     "ProviderAdapter",
+    "GreenhouseAdapter",
+    "ProviderRegistry",
+    "DEFAULT_PROVIDER_REGISTRY",
+    "build_default_provider_registry",
 ]
-